* include/olectl.h (PROPPAGEINFO): Change type of 	field cb to ULONG.

+++ winsup/w32api/include/olectl.h
@@ -309,7 +309,7 @@ typedef struct tagOCPFIPARAMS {
 	DISPID dispidInitialProperty;
 } OCPFIPARAMS,*LPOCPFIPARAMS;
 typedef struct tagPROPPAGEINFO {
-	size_t cb;
+	ULONG cb;
 	LPOLESTR pszTitle;
 	SIZE size;
 	LPOLESTR pszDocString;
